routes protected and buttons hidden

+++ routes/web.php
@@ -33,80 +33,72 @@ Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified',
 
     // PROJECTS
     Route::get('/projects', [ProjectController::class, 'index'])->name('projects.index');
-    Route::get('/projects/create', [ProjectController::class, 'create'])->name('projects.create');
-    Route::post('/projects/create', [ProjectController::class, 'store'])->name('projects.store');
+    Route::get('/projects/create', [ProjectController::class, 'create'])->name('projects.create')->middleware('admin');
+    Route::post('/projects/create', [ProjectController::class, 'store'])->name('projects.store')->middleware('admin');
     Route::get('/projects/{id}', [ProjectController::class, 'show'])->name('projects.show');
-    Route::get('/projects/{id}/edit', [ProjectController::class, 'edit'])->name('projects.edit');
-    Route::post('/projects/{id}/edit', [ProjectController::class, 'update']);
-    Route::delete('/projects/{id}', [ProjectController::class, 'destroy'])->name('projects.destroy');
+    Route::get('/projects/{id}/edit', [ProjectController::class, 'edit'])->name('projects.edit')->middleware('admin');
+    Route::post('/projects/{id}/edit', [ProjectController::class, 'update'])->middleware('admin');
+    Route::delete('/projects/{id}', [ProjectController::class, 'destroy'])->name('projects.destroy')->middleware('admin');
 
-    Route::get('/project/{project_id}/company/{company_id}/manageUsers', [ProjectController::class,'manageUsers'])->name('projects.companies.manageUsers.index');
-    Route::post('/project/{project_id}/company/{company_id}/manageUsers', [ProjectController::class,'manageUsersStore'])->name('projects.companies.manageUsers.store');
-    Route::delete('/project/{project_id}/company/{company_id}/manageUsers/{user_id}', [ProjectController::class,'manageUsersDestroy'])->name('projects.companies.manageUsers.destroy');
+    Route::get('/project/{project_id}/company/{company_id}/manageUsers', [ProjectController::class,'manageUsers'])->name('projects.companies.manageUsers.index')->middleware('admin');
+    Route::post('/project/{project_id}/company/{company_id}/manageUsers', [ProjectController::class,'manageUsersStore'])->name('projects.companies.manageUsers.store')->middleware('admin');
+    Route::delete('/project/{project_id}/company/{company_id}/manageUsers/{user_id}', [ProjectController::class,'manageUsersDestroy'])->name('projects.companies.manageUsers.destroy')->middleware('admin');
 
 
     // COMPANIES
-    Route::get('/companies', [CompanyController::class, 'index'])->name('companies.index');
-    Route::get('/companies/create', [CompanyController::class, 'create'])->name('companies.create');
-    Route::post('/companies/create', [CompanyController::class, 'store'])->name('companies.store');
-    Route::get('/companies/{id}', [CompanyController::class, 'show'])->name('companies.show');
-    Route::get('/companies/{id}/edit', [CompanyController::class, 'edit'])->name('companies.edit');
-    Route::post('/companies/{id}/edit', [CompanyController::class, 'update']);
-    Route::delete('/companies/{id}', [CompanyController::class, 'destroy'])->name('companies.destroy');
-    // FER EL POST DE MANAGE USERS
-    Route::get('/projects/{project_id}/company/{company_id}/documents', [CompanyController::class,'showDocuments'])->name('projects.companies.documents.index');
+    Route::get('/companies', [CompanyController::class, 'index'])->name('companies.index')->middleware('admin');
+    Route::get('/companies/create', [CompanyController::class, 'create'])->name('companies.create')->middleware('admin');
+    Route::post('/companies/create', [CompanyController::class, 'store'])->name('companies.store')->middleware('admin');
+    Route::get('/companies/{id}', [CompanyController::class, 'show'])->name('companies.show')->middleware('admin');
+    Route::get('/companies/{id}/edit', [CompanyController::class, 'edit'])->name('companies.edit')->middleware('admin');
+    Route::post('/companies/{id}/edit', [CompanyController::class, 'update'])->middleware('admin');
+    Route::delete('/companies/{id}', [CompanyController::class, 'destroy'])->name('companies.destroy')->middleware('admin');
+    Route::get('/projects/{project_id}/company/{company_id}/documents', [CompanyController::class,'showDocuments'])->name('projects.companies.documents.index')->middleware('admin');
 
     // REPRESENTANTS
-    Route::get('/companies/{company_id}/create-respresentant', [RepresentantController::class, 'create'])->name('representants.create');
-    Route::post('/companies/{company_id}/create-respresentant', [RepresentantController::class, 'store'])->name('representants.store');
-    Route::get('/representants/{id}', [RepresentantController::class, 'show'])->name('representants.show');
-    Route::get('/representants/{id}/edit', [RepresentantController::class, 'edit'])->name('representants.edit');
-    Route::post('/representants/{id}/edit', [RepresentantController::class, 'update']);
-    Route::delete('/representants/{id}', [RepresentantController::class, 'destroy'])->name('representants.destroy');
+    Route::get('/companies/{company_id}/create-respresentant', [RepresentantController::class, 'create'])->name('representants.create')->middleware('admin');
+    Route::post('/companies/{company_id}/create-respresentant', [RepresentantController::class, 'store'])->name('representants.store')->middleware('admin');
+    Route::get('/representants/{id}', [RepresentantController::class, 'show'])->name('representants.show')->middleware('admin');
+    Route::get('/representants/{id}/edit', [RepresentantController::class, 'edit'])->name('representants.edit')->middleware('admin');
+    Route::post('/representants/{id}/edit', [RepresentantController::class, 'update'])->middleware('admin');
+    Route::delete('/representants/{id}', [RepresentantController::class, 'destroy'])->name('representants.destroy')->middleware('admin');
 
     // USER TYPES
-    Route::get('/userTypes/create', [UserTypeController::class, 'create'])->name('userTypes.create');
-    Route::post('/userTypes/create', [UserTypeController::class, 'store'])->name('userTypes.store');
-    Route::get('/userTypes/{id}', [UserTypeController::class, 'edit'])->name('userTypes.edit');
-    Route::post('/userTypes/{id}', [UserTypeController::class, 'update']);
-    Route::delete('/userTypes/{id}', [UserTypeController::class, 'destroy'])->name('userTypes.destroy');
+    Route::get('/userTypes/create', [UserTypeController::class, 'create'])->name('userTypes.create')->middleware('admin');
+    Route::post('/userTypes/create', [UserTypeController::class, 'store'])->name('userTypes.store')->middleware('admin');
+    Route::get('/userTypes/{id}', [UserTypeController::class, 'edit'])->name('userTypes.edit')->middleware('admin');
+    Route::post('/userTypes/{id}', [UserTypeController::class, 'update'])->middleware('admin');
+    Route::delete('/userTypes/{id}', [UserTypeController::class, 'destroy'])->name('userTypes.destroy')->middleware('admin');
 
     // CONTACTS
-    Route::get('/companies/{company_id}/create-contact', [ContactController::class, 'create'])->name('contacts.create');
-    Route::post('/companies/{company_id}/create-contact', [ContactController::class, 'store'])->name('contacts.store');
-    Route::get('/companies/contacts/{id}', [ContactController::class, 'show'])->name('contacts.show');
-    Route::get('/companies/contacts/{id}/edit', [ContactController::class, 'edit'])->name('contacts.edit');
-    Route::post('/companies/contacts/{id}/edit', [ContactController::class, 'update']);
-    Route::delete('/companies/contacts/{id}', [ContactController::class, 'destroy'])->name('contacts.destroy');
+    Route::get('/companies/{company_id}/create-contact', [ContactController::class, 'create'])->name('contacts.create')->middleware('admin');
+    Route::post('/companies/{company_id}/create-contact', [ContactController::class, 'store'])->name('contacts.store')->middleware('admin');
+    Route::get('/companies/contacts/{id}', [ContactController::class, 'show'])->name('contacts.show')->middleware('admin');
+    Route::get('/companies/contacts/{id}/edit', [ContactController::class, 'edit'])->name('contacts.edit')->middleware('admin');
+    Route::post('/companies/contacts/{id}/edit', [ContactController::class, 'update'])->middleware('admin');
+    Route::delete('/companies/contacts/{id}', [ContactController::class, 'destroy'])->name('contacts.destroy')->middleware('admin');
 
     // MANAGE COMOPANIES
-    Route::get('/projects/{project_id}/manage-companies', [ProjectController::class, 'manageCompanies'])->name('projects.manageCompanies.index');
-    Route::post('/projects/{project_id}/manage-companies', [ProjectController::class, 'manageCompaniesStore'])->name('projects.manageCompanies.store');
-    Route::delete('/projects/{project_id}/manage-companies/{company_id}', [ProjectController::class, 'manageCompaniesDestroy'])->name('projects.manageCompanies.destroy');
+    Route::get('/projects/{project_id}/manage-companies', [ProjectController::class, 'manageCompanies'])->name('projects.manageCompanies.index')->middleware('admin');
+    Route::post('/projects/{project_id}/manage-companies', [ProjectController::class, 'manageCompaniesStore'])->name('projects.manageCompanies.store')->middleware('admin');
+    Route::delete('/projects/{project_id}/manage-companies/{company_id}', [ProjectController::class, 'manageCompaniesDestroy'])->name('projects.manageCompanies.destroy')->middleware('admin');
 
     // PHASES
-    Route::get('/projects/{project_id}/create-phase', [PhaseController::class, 'create'])->name('projects.phases.create');
-    Route::post('/projects/{project_id}/create-phase', [PhaseController::class, 'store'])->name('projects.phases.store');
+    Route::get('/projects/{project_id}/create-phase', [PhaseController::class, 'create'])->name('projects.phases.create')->middleware('admin');
+    Route::post('/projects/{project_id}/create-phase', [PhaseController::class, 'store'])->name('projects.phases.store')->middleware('admin');
     Route::get('/projects/{project_id}/phases/{phase_id}', [PhaseController::class, 'show'])->name('projects.phases.show');
-    Route::get('/projects/{project_id}/edit-phase/{phase_id}', [PhaseController::class, 'edit'])->name('projects.phases.edit');
-    Route::post('/projects/{project_id}/edit-phase/{phase_id}', [PhaseController::class, 'update']);
-    Route::delete('/projects/{project_id}/delete-phase/{phase_id}', [PhaseController::class, 'destroy'])->name('projects.phases.destroy');
+    Route::get('/projects/{project_id}/edit-phase/{phase_id}', [PhaseController::class, 'edit'])->name('projects.phases.edit')->middleware('admin');
+    Route::post('/projects/{project_id}/edit-phase/{phase_id}', [PhaseController::class, 'update'])->middleware('admin');
+    Route::delete('/projects/{project_id}/delete-phase/{phase_id}', [PhaseController::class, 'destroy'])->name('projects.phases.destroy')->middleware('admin');
 
     // SUBPHASES
-    Route::get('/projects/{project_id}/phases/{phase_id}/create-subphase', [SubphaseController::class, 'create'])->name('projects.phases.subphases.create');
-    Route::post('/projects/{project_id}/phases/{phase_id}/create-subphase', [SubphaseController::class, 'store'])->name('projects.phases.subphases.store');
+    Route::get('/projects/{project_id}/phases/{phase_id}/create-subphase', [SubphaseController::class, 'create'])->name('projects.phases.subphases.create')->middleware('admin');
+    Route::post('/projects/{project_id}/phases/{phase_id}/create-subphase', [SubphaseController::class, 'store'])->name('projects.phases.subphases.store')->middleware('admin');
     Route::get('/projects/{project_id}/phases/{phase_id}/subphases/{subphase_id}', [SubphaseController::class, 'show'])->name('projects.phases.subphases.show');
-    Route::get('/projects/{project_id}/phases/{phase_id}/edit-subphase/{subphase_id}', [SubphaseController::class, 'edit'])->name('projects.phases.subphases.edit');
-    Route::post('/projects/{project_id}/phases/{phase_id}/edit-subphase/{subphase_id}', [SubphaseController::class, 'update']);
-    Route::delete('/projects/{project_id}/phases/{phase_id}/delete-subphase/{subphase_parent_id}', [SubphaseController::class, 'destroy'])->name('projects.phases.subphases.destroy');
+    Route::get('/projects/{project_id}/phases/{phase_id}/edit-subphase/{subphase_id}', [SubphaseController::class, 'edit'])->name('projects.phases.subphases.edit')->middleware('admin');
+    Route::post('/projects/{project_id}/phases/{phase_id}/edit-subphase/{subphase_id}', [SubphaseController::class, 'update'])->middleware('admin');
+    Route::delete('/projects/{project_id}/phases/{phase_id}/delete-subphase/{subphase_parent_id}', [SubphaseController::class, 'destroy'])->name('projects.phases.subphases.destroy')->middleware('admin');
 
-    // SUBPHASES
-    Route::get('/projects/{project_id}/phases/{phase_id}/create-subphase', [SubphaseController::class, 'create'])->name('projects.phases.subphases.create');
-    Route::post('/projects/{project_id}/phases/{phase_id}/create-subphase', [SubphaseController::class, 'store'])->name('projects.phases.subphases.store');
-    Route::get('/projects/{project_id}/phases/{phase_id}/subphases/{subphase_id}', [SubphaseController::class, 'show'])->name('projects.phases.subphases.show');
-    Route::get('/projects/{project_id}/phases/{phase_id}/edit-subphase/{subphase_id}', [SubphaseController::class, 'edit'])->name('projects.phases.subphases.edit');
-    Route::post('/projects/{project_id}/phases/{phase_id}/edit-subphase/{subphase_id}', [SubphaseController::class, 'update'])->name('projects.phases.subphases.update');
-    Route::delete('/projects/{project_id}/phases/{phase_id}/delete-subphase/{subphase_id}', [SubphaseController::class, 'destroy'])->name('projects.phases.subphases.destroy');
 
     // DOCUMENTS
     Route::get('/projects/{project_id}/phases/{phase_id}/subphases/{subphase_id}/upload-documents', [DocumentController::class, 'create'])->name('projects.phases.subphases.document.create');
@@ -117,14 +109,19 @@ Route::middleware(['auth:sanctum', config('jetstream.auth_session'), 'verified',
     Route::get('/edit-document/{document_id}', [DocumentController::class, 'edit'])->name('document.edit');
     Route::post('/edit-document/{document_id}', [DocumentController::class, 'update'])->name('document.update');
     Route::delete('/delete-document/{document_id}', [DocumentController::class, 'destroy'])->name('document.destroy');
-    Route::get('/project/{project_id}/company/{company_id}/downloadAll', [DocumentController::class,'downloadAll'])->name('projects.companies.documents.downloadAll');
+    Route::get('/project/{project_id}/company/{company_id}/downloadAll', [DocumentController::class,'downloadAll'])->name('projects.companies.documents.downloadAll')->middleware('admin');
 
     // FEEDBACK
-    Route::get('/documents/{document_id}/feedbacks', [FeedbackController::class, 'index'])->name('feedbacks.index');
-    Route::get('/documents/{document_id}/feedbacks/create', [FeedbackController::class, 'create'])->name('feedbacks.create');
-    Route::post('/documents/{document_id}/feedbacks/create', [FeedbackController::class, 'store'])->name('feedbacks.store');
-    Route::get('/documents/{document_id}/feedbacks/{feedback_id}/edit', [FeedbackController::class, 'edit'])->name('feedbacks.edit');
-    Route::post('/documents/{document_id}/feedbacks/{feedback_id}/edit', [FeedbackController::class, 'update'])->name('feedbacks.update');
-    Route::delete('/feedbacks/{feedback_id}', [FeedbackController::class, 'destroy'])->name('feedbacks.destroy');
+    Route::get('/documents/{document_id}/feedbacks', [FeedbackController::class, 'index'])->name('feedbacks.index')->middleware('admin');
+    Route::get('/documents/{document_id}/feedbacks/create', [FeedbackController::class, 'create'])->name('feedbacks.create')->middleware('admin');
+    Route::post('/documents/{document_id}/feedbacks/create', [FeedbackController::class, 'store'])->name('feedbacks.store')->middleware('admin');
+    Route::get('/documents/{document_id}/feedbacks/{feedback_id}/edit', [FeedbackController::class, 'edit'])->name('feedbacks.edit')->middleware('admin');
+    Route::post('/documents/{document_id}/feedbacks/{feedback_id}/edit', [FeedbackController::class, 'update'])->name('feedbacks.update')->middleware('admin');
+    Route::delete('/feedbacks/{feedback_id}', [FeedbackController::class, 'destroy'])->name('feedbacks.destroy')->middleware('admin');
 
 });
+
+Route::get('{any}', function () {
+    return view('404');
+})->where('any', '.*');
+
